Update README.md
There may be more dependencies that I am missing but at least OpenG toolkit was needed to not get a broken arrow when opening the UML modeler.
Also, an end of line is needed between the git command and the xcopy command because, at least in my case, I needed to execute them separately. Added as well " marks for the paths otherwise xcopy returns an error.